% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{print.bioanchor_metadata}
\alias{print.bioanchor_metadata}
\title{Print method for bioanchor_metadata}
\usage{
\method{print}{bioanchor_metadata}(x, ...)
}
\arguments{
\item{x}{A \code{bioanchor_metadata} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print method for bioanchor_metadata
}
